gl_engine: revised the effect logic
- consolidated the effect implementation from renderer
- applied the deferred initialization
